Fix kbfs and updater bin names for linux
